requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
pandas==2.1.3
numpy==1.26.2
pyyaml==6.0.1
psutil==5.9.6
matplotlib==3.8.2
seaborn==0.13.0
locust==2.17.0
python-multipart==0.0.6
